+++ packages/testing/tsdown.config.ts
@@ -5,12 +5,20 @@ export default defineConfig({
   dts: true,
   format: ["esm", "cjs"],
   platform: "node",
-  external: [
-    "@fedify/fedify",
-    "@fedify/fedify/federation",
-    "@fedify/fedify/nodeinfo",
-    "@fedify/fedify/runtime",
-    "@fedify/fedify/vocab",
-    "@fedify/fedify/webfinger",
-  ],
+  outExtensions({ format }) {
+    return {
+      js: format === "cjs" ? ".cjs" : ".js",
+      dts: format === "cjs" ? ".d.cts" : ".d.ts",
+    };
+  },
+  deps: {
+    neverBundle: [
+      "@fedify/fedify",
+      "@fedify/fedify/federation",
+      "@fedify/fedify/nodeinfo",
+      "@fedify/fedify/runtime",
+      "@fedify/fedify/vocab",
+      "@fedify/fedify/webfinger",
+    ],
+  },
 });