+++ CHANGES.md
@@ -6,7 +6,7 @@ Fedify changelog
 Version 2.1.1
 -------------
 
-To be released.
+Released on March 27, 2026.
 
 ### @fedify/fedify